AI Roaster/requirements.txt
opencv-python
ultralytics
requests
pyyaml
